Added object hash map to check what to update
Stores a json file in .cache folder called algolia-index.json with the ids and hash of each object that gets synced to Algolia. On consequent updates (it the cache has not been cleared) it will only update and delete the changes instead of inserting everything.

This saves the amount of operations used in Algolia.

Supports multiple queries and different indexes.